requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.1.0
bleach==6.2.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
cryptography==44.0.1
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
firebase-admin==6.6.0
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
gotrue==2.11.3
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
postgrest==0.19.3
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==8.1
realtime==2.3.0
requests==2.32.3
rsa==4.9
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
storage3==0.11.3
StrEnum==0.4.15
stripe==11.5.0
supabase==2.13.0
supafunc==0.9.3
typing_extensions==4.12.2
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==14.2
yarl==1.18.3
